refactor: 파서 클래스 패키지 이동 (domain -> util)
